backend/bot/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
certifi==2025.10.5
click==8.3.0
fastapi==0.118.3
frozenlist==1.8.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
propcache==0.4.1
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==22.5
redis==6.4.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
yarl==1.22.0
passlib
python-jose
psycopg2-binary
python-multipart
python-telegram-bot
